feat: :sparkles: Added the sidebar bottom menu and refactored the sidebar
